+++ .eslintrc.cjs
@@ -25,5 +25,6 @@ module.exports = defineConfig({
         "@typescript-eslint",
     ],
     rules: {
+        "@typescript-eslint/ban-ts-comment": "off",
     },
 });