+++ src/components/MainDrawer.vue
@@ -3,12 +3,15 @@ import TabSearchVue from "./TabSearch.vue";
 import TabSettingsVue from "./TabSettings.vue";
 
 import { inject, Ref } from "vue";
-import { Status, map } from "./../utils/status";
+import { IConfig } from "./../types/config";
+import { Status, map } from "../utils/status";
 
 const visible = inject("visible") as Ref<boolean>; // 是否显示
 const status = inject("status") as Ref<Status>; // 连接状态
 const message = inject("message") as Ref<string>; // 连接状态消息
 const version = inject("version") as Ref<string>; // 内核版本
+
+const config = inject("config") as IConfig; // 用户配置
 
 function handleOk() {
     visible.value = false;
@@ -33,29 +36,43 @@ function handleCancel() {
                     src="./../assets/siyuan-32.png"
                 />
                 <!-- REF [Arco Design Vue](https://arco.design/vue/component/popover) -->
-                <a-popover position="br">
+                <!-- 鼠标悬浮气泡卡片 -->
+                <a-popover position="bl">
                     <!-- REF [Arco Design Vue](https://arco.design/vue/component/badge) -->
+                    <!-- 鼠标悬浮的元素 -->
                     <a-badge
                         class="title-label"
                         :status="status"
                         :text="$t('siyuan')"
                     />
+                    <!-- 气泡卡片标题 -->
                     <template #title>
-                        {{ $t("status_server") }}
-                        <!-- REF [Arco Design Vue](https://arco.design/vue/component/tag) -->
-                        <a-tag>
-                            <template #icon>
-                                <icon-info-circle v-show="status === Status.normal" />
-                                <icon-clock-circle v-show="status === Status.processing" />
-                                <icon-check-circle v-show="status === Status.success" />
-                                <icon-exclamation-circle v-show="status === Status.warning" />
-                                <icon-close-circle v-show="status === Status.danger" />
-                            </template>
-                            {{ version }}
-                        </a-tag>
+                        {{ $t("server_status") }}
+                        <!-- REF [Arco Design Vue](https://arco.design/vue/component/tooltip) -->
+                        <!-- 显示思源服务源的文字气泡 -->
+                        <a-tooltip
+                            :content="config.server.url.origin"
+                            position="bottom"
+                            mini
+                        >
+                            <!-- REF [Arco Design Vue](https://arco.design/vue/component/tag) -->
+                            <!-- 显示思源版本号的标签 -->
+                            <a-tag>
+                                <template #icon>
+                                    <icon-info-circle v-show="status === Status.normal" />
+                                    <icon-clock-circle v-show="status === Status.processing" />
+                                    <icon-check-circle v-show="status === Status.success" />
+                                    <icon-exclamation-circle v-show="status === Status.warning" />
+                                    <icon-close-circle v-show="status === Status.danger" />
+                                </template>
+                                {{ version }}
+                            </a-tag>
+                        </a-tooltip>
                     </template>
+                    <!-- 气泡卡片内容 -->
                     <template #content>
                         <!-- REF [Arco Design Vue](https://arco.design/vue/component/alert) -->
+                        <!-- 显示上次状态提示信息 -->
                         <a-alert
                             :title="map(status).toUpperCase()"
                             :type="(map(status) as any)"
@@ -65,6 +82,7 @@ function handleCancel() {
                     </template>
                 </a-popover>
                 <!-- REF [Arco Design Vue](https://arco.design/vue/component/input-tag) -->
+                <!-- 搜索输入框 -->
                 <a-input-tag
                     class="title-input"
                     size="mini"