+++ src/main.ts
@@ -13,40 +13,26 @@ import en from "./locales/en";
 import zh_Hans from "./locales/zh-Hans";
 import zh_Hant from "./locales/zh-Hant";
 
-const app = createApp(App);
+import { mapLocal } from "./utils/language";
 
 /* 语言包 */
 const messages = {
-    en,
-
-    zh: zh_Hans,
+    "en": en,
     "zh-Hans": zh_Hans,
-    "zh-CN": zh_Hans,
-
     "zh-Hant": zh_Hant,
-    "zh-TW": zh_Hant,
-    "zh-HK": zh_Hant,
 };
 
-let locale: string = navigator.language;
-switch (true) {
-    case locale in messages:
-        break;
-    case locale.startsWith("en"):
-        locale = "en";
-        break;
-    case locale.startsWith("zh"):
-        locale = "zh";
-        break;
-    default:
-        break;
-}
+const locale = mapLocal(navigator.language, messages);
+const fallbackLocale = "en";
 
 const i18n = createI18n({
     locale, // set locale
-    fallbackLocale: "en", // set fallback locale
+    fallbackLocale, // set fallback locale
     messages,
 });
+
+const app = createApp(App);
+app.provide("i18n", i18n); // 提供全局依赖
 
 app.use(i18n); // 国际化
 app.use(ArcoVue); // Arco 组件库